+++ src/tests/util.rs
@@ -27,7 +27,7 @@ pub async fn init_test_db() -> Database {
     .unwrap();
 
     sqlx::query(
-        "CREATE TABLE tokens(\n    id INTEGER PRIMARY KEY AUTOINCREMENT,\n    chat_id INTEGER NOT NULL,\n    token TEXT NOT NULL,\n    issued_at INTEGER NOT NULL,\n    last_used_at INTEGER,\n    revoked_at INTEGER\n)"
+        "CREATE TABLE tokens(\n    id INTEGER PRIMARY KEY AUTOINCREMENT,\n    chat_id INTEGER NOT NULL,\n    token TEXT NOT NULL,\n    name TEXT,\n    issuer_user_id INTEGER,\n    issuer_name TEXT,\n    issued_at INTEGER NOT NULL,\n    last_used_at INTEGER,\n    revoked_at INTEGER\n)"
     )
     .execute(&pool)
     .await